+++ Ecommerce_WebApi_Application/Models/DisplayCartModel.cs
@@ -17,7 +17,7 @@ namespace Ecommerce_WebApi_Application.Models
         public int? Quantity { get; set; }
 
 
-        public int? Product_Price { get; set; }
+        public decimal? Product_Price { get; set; }
